feat: Check for `CI` env var to catch unknown CI envs (#130)
This will let us know if a build is being ran in a CI system that we don't
support. Most CI systems will set the `CI` env var to true, so this can give us
a better understanding of how many folks are running in CI systems that we're
not supporting.