:recycle: Remove large deps in favor of smaller deps
Node has native promises, so bluebird is unnecessary. The bluebird-retry package
was replaced with a promise-retry package that works with native promises.

The request-promise package is another large package which was replaced by
node-fetch. Both packages do not return standard responses or errors, so
additional work was necessary to mimic request-promise's responses and errors
for backwards compatibility.

Both of these changes prompted moving the duplicate logic into a retryRequest
utility method.